requirements.txt
asgiref==3.4.1
certifi==2021.5.30
dj-database-url==0.5.0
Django==3.2.5
django-phonenumber-field==5.2.0
django-restframework==0.0.1
djangorestframework==3.12.4
gunicorn==20.1.0
phonenumbers==8.12.26
phonenumberslite==8.12.26
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
wincertstore==0.2
